fix: BNO08X_SPI_CRC_SIZE -> BNO08X_SPI_SHTP_CRC_SIZE and double-promotion in BNO08x.c:290
Fix undeclared BNO08X_SPI_CRC_SIZE in BNO08x_SPI.c (should be BNO08X_SPI_SHTP_CRC_SIZE). Fix missing float-to-double cast in BNO08x.c:290 LOG_INF.

+++ src/sensor/imu/BNO08x.c
@@ -287,7 +287,7 @@ int bno08x_init(float clock_rate, float accel_time, float gyro_time,
 
     (void)clock_rate;
 
-    LOG_INF("BNO08x init: accel %.3f s, gyro %.3f s", accel_time, gyro_time);
+    LOG_INF("BNO08x init: accel %.3f s, gyro %.3f s", (double)accel_time, (double)gyro_time);
 
     /* Wait for boot advertisement */
     uint8_t pkt_buf[BNO08X_SHTP_MAX_PACKET];
